ref(api): Move OrganizationEventsEndpointBase into its own file
