+++ src/sentry/api/bases/__init__.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import
 
 from .group import *  # NOQA
 from .organization import *  # NOQA
+from .organization_events import *  # NOQA
 from .organizationissues import *  # NOQA
 from .organizationmember import *  # NOQA
 from .project import *  # NOQA
